Updated __repr__ in Parameter class from parameter.py
After appling minimize the parameters adquire an std_err, even if they are fixed. For those parameters the __repr__ return two times the word value, for example:
<Parameter 'c', value=value=0 +/- 0 (fixed), bounds=[-inf:inf]>

I have invert the order of the "if"s and put one of them as an "else if". The bad new is that know the uncertainty information of the fixed parameters is not shown... However I don't know why it should be printed...