BUG: parameter 'name' not passed to Model class for built-in models in models.py.
The parameter 'name' was present in the __init__ function of the class,
but not updated in the kwargs that are passed to the Model class.
Removed 'name' from the __init__ so that it doesn't show up in the call
signture anymore, but is now recognized as **kwargs and is passed on.