DEP: change from 'missing' to 'nan_policy' in the call signature of built-in models.
The use of 'missing' is deprecated in lmfit v0.9.8 and 'nan_policy'
should be used; therefore, 'missing' has been removed from the call
signature. It is replaced by 'nan_policy', which is also updated in the
kwargs.update() statement.  However, if 'missing' is used that will
still work since it is recognized as **kwargs and is passed to the
Model class.