Deleted sampler_kwargs and added progress=True
The opposite of what I hoped for, I had to manually add the flag `progress=True` (`line 879`) and `progress=progress` (`line 1200`), because the progress bar is allocated in the `self.sampler` method `self.sampler.run` (where `self.sampler = EnsembleSampler(...)`).  

It works great; but the precedent is that we need to add a new flag for each `.run` kwarg.