include a emcee major version check
Because the latest PYPI version of `emcee` does not include the `progress` kwarg -- to be released with v3.0.dev0 -- we added a version check to confirm that the `emcee` major version installed is >3.  If it is, then `lmfit` with use progress with `progress=progress` on `line 1201`. The flag (`progress=True` `line 880`) can still be deactivated manually, even if `emcee.__version__ >= 3`.