FIX: numdifftools needs len(residual) > len(pars)
Some tests (test_ampgo.py, test_basinhopping.py) were failing. To me it
seems the reason is there we are trying to find the minimum of a
function instead of fitting a function to data points.

I *think* that numdifftools requires that the output of a function is
an array with more elements than fitting parameters. Adding this
requirement will not calculate a covariance_matrix when fitting a
function, but probably not many people use lmfit for this anyway...