Added progress bar to emcee call (#496)
* Add `sampler_kwargs` as kwarg to Minimizer.emcee

By adding `sampler_kwargs` as a kwarg to `Minimizer.emcee`, we can now send kwargs directly to emcee that are not consider by default in the current Minimizer instance.

Example:

`emcee` was updated to use `tqdm` as a progress bar over the MCMC sampling; but this was not added into the `lmfit-py`. To avoid requiring `lmfit-py` to update itself with every new kwarg from `emcee`, passing a `sampler_kwargs` into the `Minimizer` instance allows users to access all `emcee` kwargs directly.

* Deleted sampler_kwargs and added progress=True

The opposite of what I hoped for, I had to manually add the flag `progress=True` (`line 879`) and `progress=progress` (`line 1200`), because the progress bar is allocated in the `self.sampler` method `self.sampler.run` (where `self.sampler = EnsembleSampler(...)`).  

It works great; but the precedent is that we need to add a new flag for each `.run` kwarg.

* minimized the number of PR corrections

* add tqdm to the dependencies

* updated requirements for emcee

* removed dependencies for sub features

* include a emcee major version check

Because the latest PYPI version of `emcee` does not include the `progress` kwarg -- to be released with v3.0.dev0 -- we added a version check to confirm that the `emcee` major version installed is >3.  If it is, then `lmfit` with use progress with `progress=progress` on `line 1201`. The flag (`progress=True` `line 880`) can still be deactivated manually, even if `emcee.__version__ >= 3`.

* Include emcee v3 check in self._lastpos line1200

* Configured EMCEE_VERSION Check and output

Changes the variable name and behaviour of the EMCEE version check for version>=3, now it runs one set of 2 lines, as oppose to the default set of 2 lines