REL: explicitly add install_requires in setup.py
Currently, the requirements do not end up in the json meta-data on PyPI
which limits the usability of tools that rely on it. For example, the
tool upt (https://framagit.org/upt/upt) that we are currently exploring
to help with packaging for MacPorts does not report any dependencies.

Apparently, the information is computed locally when generating the
PyPI package and, therefore, might depend on how this is exactly done.
Nevertheless, explicitly listing the dependencies should resolve this.