FIX: only check if a parameter is at its initial value if an initial value is present
Probably this is a rare occurence as most users will actually set an
initial value when initializing the parameters. However, for algorithms
like "brute" that is actually not required and this might happen.

Added a regression test and updated the code to include the check
whether an init_value is present.

See: #578