DOC: add examples from documentation to gallery
For the gallery one needs a correct header/docstring for the conversion
to work. We don't want this for our examples that are part of the
documentation where we often just show the file content.

So we copy the doc_*.py examples to a separate "documentation" folder
and add a header. In this way they will also show up in a "sub-gallery".