DOC: add comment about the failure of doc_model_loadresult.py
For some reason the conversion by sphinx-gallery for
doc_model_loadresult.py fails, even though execute the code in a Python
shell or Jupyter notebook works fine.

Error raised = NameError: name 'np' is not defined

Add a note to the resulting example in the gallery and add this file to
'expected_failing_examples' to make sure the build finishes.