ENH: improve message if ValueError is raised by _nan_policy
The original message ("ValueError: The input contains nan values")
seems often -incorrectly-  imply to users that their input data contains
NaN values. The "input" here is referring to the input data of the
_nan_policy function.

Clarify that it can either be caused by the input data or, more likely,
by the output of the objective/model function. Additionally, include a
link to the FAQ on the documentation website.