TST: Add codecov configuration
Relax the coverage requirements a bit, to make sure that PRs do not
fail when coverage changes are very small. With this PR we allow:
- a 0.5% drop in overall coverage
- a 10% missed coverage in the patch

Of course, the goal should still be to increase the code coverage with
every commit (or at least keep it the same), and attempt to have all
your changes covered with test.