FIX: reset Minimizer()._abort to False for each fit
If a fit is aborted (e.g., by an iter_cb function) the ._abort attribute
of the Minimizer class is set to True. Performing another fit using the
same Minimizer instance will now fail, because of this.

Since every method of doing a fit and every algorithm calls
"prepare_fit" at the start, this seems the best place to resolve the
issue.

Closes: #756, #757